requirements.txt
paho_mqtt==1.6.1
python_telegram_bot==13.11